PROJECT-3/secure-file-sharing-system/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.2.1
cryptography==45.0.4
Flask==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pycparser==2.22
pycryptodome==3.23.0
python-dotenv==1.1.1
Werkzeug==3.1.3
